DEFAULT table id fields to uuid_generate_v4()
Updated create table statements to add DEFAULT uuid_generate_v4() which produces a uuid for each id in each table if needed. 

Added NOT NULL statement to the oname table for the naming_authority and obj fields to prevent entry of names without proper foreign key connections. Other tables should be reviewed for the same issue.